+++ Services/Data/SongSerializer.cs
@@ -55,7 +55,7 @@ namespace Maestro.Services.Data
             if (dto.Notes != null)
             {
                 song.Notes.AddRange(dto.Notes);
-                var commands = NoteParser.Parse(dto.Notes);
+                var commands = SongCompiler.Parse(dto.Notes, song.Instrument);
                 song.Commands.AddRange(commands);
             }
 
@@ -86,7 +86,7 @@ namespace Maestro.Services.Data
                 if (dto.Notes != null)
                 {
                     song.Notes.AddRange(dto.Notes);
-                    var commands = NoteParser.Parse(dto.Notes);
+                    var commands = SongCompiler.Parse(dto.Notes, song.Instrument);
                     song.Commands.AddRange(commands);
                 }